master changes in mmfsl committed.
git-svn-id: http://192.168.110.32/svn/Products/PFF/git-migration/branches/mlod@120685 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ pff-model/src/main/java/com/pennanttech/pff/mmfl/cd/model/MerchantDetails.java
@@ -11,7 +11,7 @@ public class MerchantDetails extends AbstractWorkflowEntity {
 
 	private long merchantId = Long.MIN_VALUE;
 	private String merchantName;
-	private int storeId;
+	private BigDecimal storeId;
 	private String storeName;
 	private String storeAddressLine1;
 	private String storeAddressLine2;
@@ -53,11 +53,11 @@ public class MerchantDetails extends AbstractWorkflowEntity {
 		this.merchantName = merchantName;
 	}
 
-	public int getStoreId() {
+	public BigDecimal getStoreId() {
 		return storeId;
 	}
 
-	public void setStoreId(int storeId) {
+	public void setStoreId(BigDecimal storeId) {
 		this.storeId = storeId;
 	}
